requirements.txt
cn2an
einops
ffmpeg_python
g2p_en
gruut
jieba_fast
LangSegment
librosa
matplotlib
numpy
pandas
pyopenjtalk
pypinyin
PySide6
pytorch_lightning
PyYAML
regex
soundfile
torch
torchmetrics
tqdm
transformers
typeguard
